test(integration): extract resilience chat helpers
